Added packages and files for the experiment home page and its tabs
I'm currently thinking i'll rework the experiment activity into a fragment which will have TabViews for each of the options on the page, but i'm still working stuff out. This would fix the problem I was having with how to overlay it on the Main Activity as it would just work like the other Home fragments, but with more going on inside it. and we already know how to pass data, like experiment ID and user ID to a fragment so it also fixes that issue